SystemUI needs the SET_WALLPAPER permission
If changing wallpaper when low on memory, retrieving the default
wallpaper may return null, which triggers error handling in
ImageWallpaper. This error handling tries to perform
WallpaperManager.clear(), but for that it needs SET_WALLPAPER
permission, which it does not have.

For users with apps that auto change wallpaper, this issue can
be frequent in low-memory conditions.

The solution is to add the permisson.

Change-Id: I81503c1667e3952c2dd15599969f7dcc51623e5b
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>