Kill process asynchronously on updateOomAdjLocked.
Kill process takes time, especially on killProcessGroup. If it is called
by updateOomAdjLocked, it keeps lock of mServices for moments and other
requests would be delayed, like attachApplication.

If is called frequently, it may occur latency problem. Killing job is
needed to be done asynchronously in updateOomAdjLocked at least.

Change-Id: I745851943e8dbee7f823a8b4aa17c521e43e2e06

Conflicts:
	services/core/java/com/android/server/am/ActivityManagerService.java