Observer DEV_FORCE_SHOW_NAVBAR in LockTaskNotify
Since a new WindowManagerPolicy is being created in LockTaskNotify,
the init() method is never called and therefore the SettingsObserver
in PhoneWindowManager is never created and started.  This causes
hasNavigationBar() to always return false when hardware keys are
disabled.

This patch adds an observer in LockTaskNotify so we can observe this
change and correctly display the appropriate toast messages when an
app is pinned/locked.

Change-Id: I917bf37cc51c8fa587ae07a8b4a195a3d147d59e
REF: BACON-3826

Conflicts:
	services/core/java/com/android/server/am/LockTaskNotify.java