Keyguard: Don't show keyguard if CryptKeeper enabled
Some devices are exhibiting an issue where the keyguard will be
shown on top of the decrypt device screen when they have their
sim locked.  When this occurs, the user is not able to enter their
pin/password to decrypt the device even after the keyguard is
dismissed.

REF: NIGHTLIES-1254

Conflicts:
	packages/SystemUI/src/com/android/systemui/keyguard/KeyguardViewMediator.java
Change-Id: I7362b77a6ea2ec41c8835ae763ce427c12c02138

Conflicts:
	packages/SystemUI/src/com/android/systemui/keyguard/KeyguardViewMediator.java